Move shared components to shared directory
Relocated Footer, GetInTouchSection, and Header components to src/components/shared for better organization. Updated imports in about-us and main page files to reflect the new paths.
