+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Carousel from '@/components/home/carousel';
-import Footer from '@/components/footer';
-import GetInTouchSection from '@/components/get-in-touch-section';
-import Header from '@/components/header';
+import Footer from '@/components/shared/footer';
+import GetInTouchSection from '@/components/shared/get-in-touch-section';
+import Header from '@/components/shared/header';
 import ITServicesOverview from '@/components/home/it-services-overview';
 import ServicesSection from '@/components/home/services-section';
 import ScriptLoader from '@/components/script-loader';
